refactor: переместить HomePage в отдельную директорию
